turn get creds into a util function'
